[fix] unable to extract dr. ziegler skin when using its name
Issue is caused by the Unlock.GetName() function stripping dots from the filename and it not being removed in other places.

Not too sure what the reason was initially to strip dots from the name? Dots are valid in file and folder names?

If it is a issue with filenames basically everthing that gets saved now goes through a GetValidFilename function, it can be added there if it needs to.