tankview: improve image grid view performance slightly
Changes it so scrolling only scrolls 2 rows at a time (defaults to 3) and makes it so it no longer loads other images that are outside your viewport as there is no point. Loading images is thread blocking anyway so there is no gains to loading images ahead of you as when you scroll it will load more images ahead of those ones anyway.